ganti type data float ke double

+++ database/migrations/2024_05_16_025310_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('nama_produk');
             $table->foreignId('product_category_id');
-            $table->float('harga');
+            $table->double('harga');
             $table->integer('stok');
             $table->timestamps();
         });
