+++ database/migrations/2024_05_16_025343_create_transaction_details_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignId('transaction_id');
             $table->foreignId('product_id');
             $table->integer('jumlah');
-            $table->float('subtotal');
+            $table->double('subtotal');
             $table->timestamps();
         });
     }
